backend/requirements.txt
fastapi
uvicorn
python-multipart
# LangChain & AI
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-google-genai
sentence-transformers
faiss-cpu
# PDF Parsing
pypdf
# Utils
python-dotenv
httpx
